refactor(bp): move entry form to Entries page; seed demo BP data
Splits BloodPressureSection in two:
- BloodPressureForm — entry form + recent-readings list, lives on
  EntriesPage alongside Bloodwork / Form check / DNA. Where logging
  primitives belong.
- BloodPressureSection — chart + full history (delete still allowed),
  stays on Orient as a read-only trend view.

Adds seed_blood_pressure to seed_demo.py: ~30 morning + evening
readings spread across the 90-day demo window so the Orient chart
renders something in preview deploys.

Adds .longevity-form-success and .longevity-entry-list-title CSS for
the new "Reading saved." confirmation and "Recent readings" header.

Updates e2e spec to drive the form from Entries (the asserted "Reading
saved." text replaces the previous heading-presence check).
