+++ AGENTS.md
@@ -58,7 +58,7 @@ python3 -c "import sqlite3; c=sqlite3.connect('vitalscope.db'); c.row_factory=sq
   - `water_intake(id PK, date, time, amount_ml, created_at)` — per-drink entries; aggregate via `GET /api/water/daily` (`SUM(amount_ml) GROUP BY date`).
   - Daily nutrient totals via `GET /api/nutrition/daily` return `[{ date, totals: { key: amount } }]` — shape is deliberately keyed by nutrient so a chart can pluck one key without client-side grouping. When adding a chart, import `fetchNutritionDaily`/`fetchWaterDaily` directly rather than shoehorning into `useMetricData` (the shape doesn't match `fetchMetric`'s generic `T`).
 - **Protocols**: three tables. `protocols(id PK, name, category IN ('drug','peptide','ped','supplement_stack','hormesis','fasting','training_block'), dose, unit, cadence, start_date, end_date, notes, created_at, time_of_day IN ('morning','noon','evening') OR NULL, recurrence_type IN ('daily','days_of_week','every_n_days','weekly_on','as_needed') NOT NULL DEFAULT 'daily', recurrence_days, recurrence_n, recurrence_anchor_date)` — one row per intervention protocol; structured recurrence fields are authoritative for scheduling, `cadence` is a free-text fallback display. `protocol_events(id PK, protocol_id FK, date, time, dose, duration_minutes, notes, created_at)` — actual logged occurrences, `ON DELETE CASCADE` from protocols. Indexes on `date` and `protocol_id`. `protocol_adherence(date, protocol_id, time_of_day, taken, PK(date, protocol_id, time_of_day))` — daily check-off mirroring supplements; `time_of_day` is stored as `''` (empty string) when the protocol has no fixed slot, never NULL, so `ON CONFLICT` upsert via the PK works. `GET /api/protocols/{date}/scheduled` expands `recurrence_type` in Python via `_protocol_scheduled_for_date` (`daily` always, `days_of_week` matches weekday, `every_n_days` checks `(date - anchor) % n == 0`, `weekly_on` matches first day in `recurrence_days`, `as_needed` never auto-scheduled) and joins `protocol_adherence` with `COALESCE(taken, 0)` so missing rows default to unchecked. `POST /api/protocols/{date}/adherence` body `{ protocol_id, time_of_day, taken }` upserts a single slot. The schema migration in `ensure_protocol_tables` does a table-rebuild when the new columns are missing (SQLite can't ALTER constraints) and defaults existing rows to `recurrence_type='daily'`. For hormesis quick-logs (Zone 2, Sauna, Cold Plunge) the backend's `find-or-create` logic lives in the **frontend** (`findOrCreateQuick` in `ProtocolsSection.tsx`), not the backend — keep it there. The "Scheduled today" check-off card in `TodayDashboard.tsx` (Act → Today) consumes the scheduled+adherence endpoints and toggles via `saveProtocolAdherence`.
-- **Blood pressure**: `blood_pressure_entries(id PK, date, time, systolic_mmhg, diastolic_mmhg, pulse_bpm, notes, created_at)` — manual entry only (no sync source). `GET/POST /api/blood-pressure`, `DELETE /api/blood-pressure/{id}`. POST validates `systolic > diastolic` and returns 400 otherwise. Multiple readings per date are allowed (no upsert). Frontend: `BloodPressureSection.tsx` wired into `OrientPage` as the "Blood pressure" section (separate from `LongevitySection`); the latest reading also surfaces as an `AgeBadge` card in `TodayMetrics.tsx` (Observe → Today's metrics). AI context category `blood_pressure` (Health group) is registered but no prompt-builder consumes it yet — wire it through `_gather_*` helpers when an analysis depends on BP.
+- **Blood pressure**: `blood_pressure_entries(id PK, date, time, systolic_mmhg, diastolic_mmhg, pulse_bpm, notes, created_at)` — manual entry only (no sync source). `GET/POST /api/blood-pressure`, `DELETE /api/blood-pressure/{id}`. POST validates `systolic > diastolic` and returns 400 otherwise. Multiple readings per date are allowed (no upsert). Frontend split: `BloodPressureForm.tsx` is the entry form (lives on `EntriesPage` under "Blood pressure" alongside Bloodwork / Form check / DNA), and `BloodPressureSection.tsx` is the read-only chart + history view on `OrientPage` (deletion is still allowed there). The latest reading also surfaces as an `AgeBadge` card in `TodayMetrics.tsx` (Observe → Today's metrics). Demo data is seeded by `seed_blood_pressure` in `seed_demo.py` (~30 readings spread across the 90-day demo window) so the Orient chart renders something in preview deploys. AI context category `blood_pressure` (Health group) is registered but no prompt-builder consumes it yet — wire it through `_gather_*` helpers when an analysis depends on BP.
 - **Bloodwork**: two tables. `bloodwork_panels(id PK, date, source, source_upload_id, lab_name, notes, confidence, created_at)` — one row per lab report. `bloodwork_results(id PK, panel_id FK, analyte, value, value_text, unit, reference_low, reference_high, reference_text, flag, sort_order)` with `ON DELETE CASCADE` from panels. Values use `value` for numbers, `value_text` for qualitative results (`"Negative"` etc.) — never both. Reference ranges use the low/high pair for numeric ranges, `reference_text` for anything else. The AI parses PDFs only on the **anthropic** provider (OpenAI chat-completions can't ingest PDFs); `OpenAIProvider.analyze_with_tool` raises 400 on `application/pdf` mime.
 - **Longevity markers**: two additional tables.
   - `biological_age_entries(id PK, date, clock_name, value, chronological_age, rate_of_ageing, notes, created_at)` — manual entry of epigenetic clock outputs (GrimAge, Horvath, PhenoAge, DunedinPACE, TelomereLength). `GET/POST /api/biological-age`, `DELETE /api/biological-age/{id}`. No upsert — multiple readings per date and clock are allowed.
